Update main.go
This harmlessly adjust some of the code in this repo which represents the demo app running in our demos. In reality, nothing is running based on this repo, it is merely a repo we can automatically create new commits against and link to from Observe demos.

Thus this change helps alter the story of the Product Catalog feature flip into something that may make sense with legit product releases.

In total, all for show - nothing functional here.